% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\docType{class}
\name{circleS4-class}
\alias{circleS4-class}
\title{An S4 class representing a circle}
\value{
An object of class circleS4.
}
\description{
An S4 class representing a circle
}
\section{Slots}{

\describe{
\item{\code{radius}}{A numeric scalar.}

\item{\code{area}}{A numeric scalar.}

\item{\code{circumference}}{A numeric scalar.}

\item{\code{color}}{A character string.}
}}

\examples{
new("circleS4", radius = 3, area = pi * 3 ^ 2, circumference = 10000000, color = "blue")
}
