% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\name{circleS4}
\alias{circleS4}
\title{Constructor function for circleS4 class.}
\usage{
circleS4(radius, color = "blue")
}
\arguments{
\item{radius}{a numeric scalar denoting the radius of the circle.}

\item{color}{a character string denoting the color of the circle.}
}
\value{
an object of class circleS4.
}
\description{
Constructor function for circleS4 class.
}
\examples{
circleS4(10, "lime green")
}
