% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df1.R
\docType{data}
\name{df1}
\alias{df1}
\title{A toy data example to illustrate shipping data (for use with your package)}
\format{A data frame with 100 rows and 1 column:
\describe{
  \item{x}{rnorm(100, 0, 1)}
}}
\source{
Spencer Lourens made it with rnorm()
}
\usage{
df1
}
\description{
A dataset containing 100 N(0, 1) variates
}
\author{
Spencer Lourens \email{slourens@iu.edu}
}
\keyword{datasets}
