% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathFuncs.R
\name{dist2}
\alias{dist2}
\title{Calculate the distance of x and y vectors (recycled if necessary)}
\usage{
dist2(x, y)
}
\arguments{
\item{x}{numeric scalar or vector}

\item{y}{numeric scalar or vector}
}
\value{
a numeric scalar denoting the Euclidian distance between x and y
}
\description{
This function calculates the usual Euclidian distance between two vectors. The
values will be recycled, if necessary.
}
\examples{
dist2(1:5, 1:4)
dist2(rnorm(5, 1, 1), rnorm(5, 0, 1))
}

