% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\docType{methods}
\name{print,circleS4-method}
\alias{print,circleS4-method}
\title{print method for circleS4 objects}
\usage{
\S4method{print}{circleS4}(x, ...)
}
\arguments{
\item{x}{An object of class circleS4}

\item{...}{ignored for this function}
}
\value{
a textual representation of the circleS4 object
}
\description{
print method for circleS4 objects
}
\examples{
print(circleS4(3, 'red'))
}
