% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\docType{methods}
\name{print,rectS4-method}
\alias{print,rectS4-method}
\title{print method for objects of class rectS4}
\usage{
\S4method{print}{rectS4}(x, ...)
}
\arguments{
\item{x}{an object of class rectS4}

\item{...}{ignored for this function}
}
\value{
a textual representation of the rectS4 object
}
\description{
print method for objects of class rectS4
}
\examples{
print(rectS4(4, 4, "blue"))
}
