% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\docType{methods}
\name{print,shapeS4-method}
\alias{print,shapeS4-method}
\title{Print method for object of class shapeS4}
\usage{
\S4method{print}{shapeS4}(x, ...)
}
\arguments{
\item{x}{an object of class shapeS4}

\item{...}{ignored for this function}
}
\value{
a textual representation of the shapeS4 object
}
\description{
Print method for object of class shapeS4
}
\examples{
print(shapeS4("red"))
}
