% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\docType{methods}
\name{print,triangleS4-method}
\alias{print,triangleS4-method}
\title{Print method for object of class triangleS4}
\usage{
\S4method{print}{triangleS4}(x, ...)
}
\arguments{
\item{x}{an object of class triangleS4}

\item{...}{ignored for this function}
}
\value{
a textual representation of the triangleS4 object
}
\description{
Print method for object of class triangleS4
}
\examples{
print(triangleS4("red", "right"))
}
