% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\docType{class}
\name{rectS4-class}
\alias{rectS4-class}
\title{Constructor function for rectS4 class.}
\value{
an object of class rectS4
}
\description{
Constructor function for rectS4 class.
}
\section{Slots}{

\describe{
\item{\code{height}}{numeric scalar representing rectangle height.}

\item{\code{width}}{numeric scalar representing rectangle width.}

\item{\code{area}}{numeric scalar representing rectangle area.}

\item{\code{perimeter}}{numeric scalar representing rectangle perimeter.}

\item{\code{color}}{character string representing rectangle color.}
}}

\examples{
new("rectS4", height = 3, width = 3, area = 9, perimeter = 10005, color = "blue")
}
