% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\name{rectS4}
\alias{rectS4}
\title{Constructor function for rectS4 class.}
\usage{
rectS4(height, width, color = "blue")
}
\arguments{
\item{height}{a numeric scalar denoting the height of the rectangle}

\item{width}{a numeric scalar denoting the width of the rectangle}

\item{color}{a character string denoting the color of the rectangle}
}
\value{
an object of class rectS4
}
\description{
Constructor function for rectS4 class.
}
\examples{
rectS4(5, 3.2, "orange")
}
