% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathFuncs.R
\docType{class}
\name{shapeS4-class}
\alias{shapeS4-class}
\title{An S4 class representing a shape}
\value{
An object of class shapeS4
}
\description{
An S4 class representing a shape
}
\section{Slots}{

\describe{
\item{\code{color}}{a character string denoting the color of the shape}
}}
\examples{
new("shapeS4", "red")
}

