% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\name{shapeS4}
\alias{shapeS4}
\title{Constructor function for shapeS4 class.}
\usage{
shapeS4(color = "blue")
}
\arguments{
\item{color}{a character string denoting the color of the shape.}
}
\value{
an object of class shapeS4.
}
\description{
Constructor function for shapeS4 class.
}
\examples{
shapeS4("purple")
}
