% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathFuncs.R
\name{sum2}
\alias{sum2}
\title{Example toy sum function}
\usage{
sum2(x, y)
}
\arguments{
\item{x, }{y numeric vectors or scalars}
}
\value{
a numeric scalar or vector denoting the sum of x and y
}
\description{
This function calculates the sum of x and y, recycled if necessary. Just a toy
function to show creation and documentation of a new function.
}
\examples{
sum2(2, 3)
sum2(1:4, 1:7)
\dontrun{
sum2("x", "y")
}
}

