% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathFuncs.R
\name{triangleS3}
\alias{triangleS3}
\title{Constructor for S3 triangle class}
\usage{
triangleS3(color = "blue", type = "right")
}
\arguments{
\item{color}{a string variable denoting the color of the triangle}

\item{type}{a string variable denoting the type of the triangle, right, isosceles, etc.}

\item{width}{a numeric scalar specifying the width of the triangle}

\item{height}{a numeric scalar specifying the height of the triangle}
}
\value{
an object of class triangleS3
}
\description{
Constructor for S3 triangle class
}
\examples{
myTriangle <- triangleS3("green", "isosceles")
}

