% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\docType{class}
\name{triangleS4-class}
\alias{triangleS4-class}
\title{An S4 class representing a triangle}
\value{
An object of class triangleS4
}
\description{
An S4 class representing a triangle
}
\section{Slots}{

\describe{
\item{\code{color}}{a character string denoting the color of the triangle}

\item{\code{type}}{a character string denoting the type of triangle, i.e. right, isosceles}
}}

\examples{
new("triangleS4", color = "red", type = "equilateral")
}
