% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Shapes.R
\name{triangleS4}
\alias{triangleS4}
\title{Constructor function for triangleS4 class.}
\usage{
triangleS4(color = "blue", type = "right")
}
\arguments{
\item{color}{character string denoting the color of the triangle.}

\item{type}{character string denoting the type of triangle.}
}
\value{
an object of class triangleS4
}
\description{
Constructor function for triangleS4 class.
}
\examples{
triangleS4("red", type = "equilateral")
}
